incident_response_system/requirements.txt
altair==5.4.1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bitarray==2.9.3
bitstring==4.2.3
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
esptool==4.8.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQL==1.6.0
Flask-Session==0.8.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
mysql-connector-python==9.2.0
narwhals==1.26.0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==6.0.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
PyPDF2==3.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reedsolo==1.7.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
selenium==4.26.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
